requirements.txt
allure-pytest==2.6.0
allure-python-commons==2.6.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
DBUtils==2.0.2
idna==3.4
importlib-metadata==4.8.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyMySQL==0.9.3
pyparsing==3.0.9
pytest==7.0.1
requests==2.27.1
six==1.16.0
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.12
zipp==3.6.0
